snop/Flask-Firebase/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Limiter==3.5.0
firebase-admin==6.6.0
python-dotenv==1.0.1
gunicorn==23.0.0
openai-whisper==20231117
torch==2.9.1
ollama==0.6.1
requests==2.32.5
transformers==4.36.0
Pillow==10.1.0
